Uses ssh authentication with hardcoded keys.
Now, the keys will have to be read from the block device. The private
key length is variable, so I'll have to fread byte by byte until I get a
0, unless I encode the length in 4 bytes. I should do that, even though
it adds some complexity. Having a single fread is definitely better.

I *still* need the type byte, because the key can be e.g. RSA or DSA,
and the public key import needs this information.