Moves hermes-service to global tokens.
The resulting code is actually *much* easier to read, without having all
this back-and-forth between service and PAM module.

There's also a lot less allocations needed, more stuff can be put on the
stack.

Overall, this architecture is also way easier to reason about. Adding
one-time tokens is going to be a breeze.