Adds with-retries support.
This leaves some time for the OS to properly mount the usb stick. The
value can and should of course be tweaked, if e.g. 5s turns out to be
too slow or too fast.

The with-retries macro takes 2 arguments: :times and :every. It will
call the body x times every y second(s) as long as the body doesn't
return a truthy value. This means that for e.g.:

(with-retries (:times 5 :every 1)
  ;; Lasts 4 seconds max
  &body)

Will be called 5 times every second, **so it will take 4 seconds max**,
if the body always return NIL.